requirements.txt
requests
attrs
pytest
lxml
allure-pytest
allure-python-commons
selenium
pytest-html
webdriver-manager